Issue #5170: Refactorings: Extract UpdateGUI function and pull up specification
The commit merges the GUI update functions (started with 'SetControls...') to a
single function. The same function is called when the reference or the working
data has changed. This simplifies the API and GUI updates are quick, so we can
do not loose anything if we update all the GUI controls after either the reference
of the working data has changed.

This commit also adds the 'const' specifier to several functions, and the return
type of GetReferenceImage() was also made a const pointer. This enforces compile time
check and prohibits modifications of the reference image.