Issue #5170: Remove setting and retrieving obsolete segmentation node properties
A segmentation node property had been set from the GUI and the GUI was
updated again from it. I do not see what was the purpose of this.

Anyway, we store the complete GUI state in segmentation node properties
now, and we can restore the complete GUI state from the properties.

Therefore, the using the old properties is remove by this commit.

This commit also has some code simplifications.