Issue #5173: Do not go 'diagonally' from 'border' pixel to 'inside' pixel
The region growing algorithm was leaking when you deleted a part of the contour and redrawn it by hand,
starting and ending the hand drawn contour exactly at the end of the ends of the broken contour segments.

If there was just a little (1px) overlap between the broken contour and the hand drawn one, it was not
leaking.

This commit supposed to resolve this issue.

From the two changes the first is just a simplification of the logic expression, a condition has been
removed that can never fulfil.

The second change is the actual fix.