Issue #5170: Do not store initial seeds in working nodes
The initial segmentation and the initial seeds were stored in a working data node,
so that it can be restored when the user presses Cancel.

However, the seeds were generated based on the segmentation after the segmentation
was started (Start/restart segmentation button was clicked), so it is not necessary
to store it, it can be calculated again. Detecting a seed per region is quick.
Moreover, it is done only for one slice at a time.

Therefore, storing the initial seeds in the working data nodes is a redundant
information, and this commit removes that.