Issue #5203: Activate DnD Display editor part when selected window changed
The DnD Display (niftk::MultiViewerEditor) is listening to the 'WindowSelected'  events
to know when the selected viewer has changed. When the selected viewer has changed, it
restores the data manager selection to the set of nodes that were selected in that
viewer, and notifies the other workbench parts.

However, the notification happens only if the DnD Display is active, but when this
listener ('OnWindowSelected') is called, it is not always. This is because you can
switch between viewers also by key interactions, and the focused renderer can still
receive these interactions, even if another workbench part is active.

So, we need to activate the DnD Display after a window is selected, so that we
do not just update the data manager (and own) selection, but can, indeed, notify
other parts of the change.