Issue #5170: Do not try to deactivate tools if no tool is active
Extra carefullness. After checking the code of mitk::ToolManager::ActivateTool(int),
I am not certain if deactivating the tools (passing -1) when no tool is active,
is harmless. There should be an equality check in this case.

I do this check before trying to deactivate the tools, even if I have not noticed
any problem with deactivating multiple times. Just to be on the safe side.