Issue #5136: Add/remove DnD Display renderers to/from rendering manager as needed
Until now all the renderers of the DnD Display and Side Viewer were registered to the
rendering manager, always. This could make the application a bit slow, especially if
many viewers were open in the DnD Display, e.g. in slice snapshot mode.

This was because when the rendering manager was requested to update all the windows,
it updated the renderers even if they were not visible on the screen.

Two attempts were made to work around this problems. The first is to set the size of
the VTK render windows to (0, 0) if they are not on the screen. This is a dirty trick.
VTK checks the size, and does not re-render the sceen if the window is zero sized.

The second is to manage the visibility of the 3D renderer differently than the 2D
renderers. Obviously, the 3D rendering is the slowest. There was a flag to control
if you wanted to see the 3D scene in the 3D window in 2x2 (aka. ortho) mode. There
used to be a check box as well on the drop down control panel or the preferences
for this.

This check box has been long ago, because I introduced different window layouts,
one with and one without 3D rendering: '2x2' and '2x2 no 3D'. For some reason,
the flag was not updated correctly, and the 3D window was empty in '2x2' mode as
well.

I also noticed, that in spite of setting the VTK render window size to 0 and
using the '2x2 no 3D' mode, the 3D VTK mapper is called, anyway. It probably
returns faster if there is no node to render. However, it also means that the
sagittal window is re-rendered if only the axial window is visible and so on.

So, to prevent all these issues happenning, to simplify the API, to avoid weird
messages from Cocoa on Mac and to make the application a bit faster, this commit
adds and removes the render windows to/from the rendering manager, as they
appear or disappear on the screen.

This commit also makes the default 'multiple window' layout to '2x2 no 3D'
instead of '2x2'. For a general purpose viewer '2x2' might be better, but
if you only want to do segmentation, you probably do not want the 3D view
at all. Also, the initialisation of the 3D window is broken at the moment,
and the image is zoomed in very much after you open it.