Issue #5129: Fixes for new PointSetDataInteractor state machine
There have been a few problems.

First, it was not possible to remove seeds by wiping, i.e. by moving the mouse
around while keeping the middle mouse button pressed.

Second, the usual locking/unlocking mechanism was not applied. I introduced this
technique to prevent interfering interactions, e.g. switching on an interactor
in the middle of doing something with the mouse in another. See this class for
more detailed explanation:

  niftk::InteractionEventObserverMutex

Finally, there were some minor but disturbing issues, e.g. the selection (diamond)
was not always removed from a seed when the mouse pointer was not over it.

This commit fixes these issues, and adds some nice features as well.

When adding a point, you can move the point by moving the mouse, until you release
the mouse button.

You can remove seeds also by shift+left mouse button, not only by middle mouse button.
This is useful if you are using a touchpad or if your mouse does not have a middle
button.

You can move seeds by dragging them with the left mouse button. This was possible
with the previous (non-data) interactor as well, but the diamond shaped cursor did not
appear over the seed, so it was not obvious when can you start dragging the point.