Issue #5187: MITK Segmentation plugin physically removed from the sources
This was a copy of an old version of the MITK Segmentation plugin in which some very instable
features (e.g. 3D interpolation of regions) had been disabled. The MITK Segmentation plugin
has improved a lot since this restricted version was made, and the original plugin is suitable
for work now. So, it is not necessary any more to keep this obsolete copy here.

Moreover, this plugin has been disabled for long time, and has not even been compiled regularly.