Issue #5213: Run region growing when changing slice with thresholding on
If you tried to remove a region by removing its seed when the thresholding was
enabled, but you did not press clean or apply just changed slice, the region was
not removed and a new seed was put back to it when you went back to the same slice
again.

Running the region growing manually with the chosen thresholds seems to resolve
the issue. I am not sure that this works well with the undo/redo functionality,
though.